Moved signals to appropriate module
